Test1

+++ common-events/src/main/java/com/myexampleproject/common/event/CartLineItem.java
@@ -4,11 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class CartLineItem {
     private String skuCode;
     private int quantity;
-    private double price;
+    private BigDecimal price;
 }
